requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
langgraph==0.0.15
typing-extensions>=4.5.0
pyyaml>=6.0
colorama>=0.4.4
tqdm>=4.65.0
pytest>=7.3.1
black>=23.3.0